dpor comments
